Added Assimp through git and loaded a model
Assimp loads and has no linker errors
Model loads, but need to check through implementing conversion to vertices this code can draw

+++ 3D Rendering Library/src/Engine/Mesh/Mesh.h
@@ -36,7 +36,7 @@ public:
 
 	void bindVBO(VBO vbo);
 
-	void renderMesh(int data_points = 12);
+	void renderMesh(size_t data_points = 12);
 
 private:
